requirements.txt
fastapi
uvicorn[standard]
asyncpg
SQLAlchemy[asyncio]
python-dotenv
openai